requirements.txt
Django==2.2.17
djangorestframework==3.12.2
python-dotenv==0.15.0
requests==2.25.1
